OPAL nest feature detection
Initialize the nest pmu and load the nest meta-data. The
nest events supported by HW are available as a lid (meta-data),
called "Catalog". The catalog lid loaded in two parts. First,
it is queued for loading in the preload function and in the second, loaded
catalog is verified and used to detect chip nest instrumentation support.
New file called "nest.c" added to "hw/" directory to contain nest support code.
Code added to include the file in Makefile.

Signed-off-by: Madhavan Srinivasan <maddy@linux.vnet.ibm.com>