skiboot: Add chip id to HOMER reserved region
This patch adds a "ibm,chip-id" property to HOMER reserved region. That
helps us in relating the HOMER reserved region to the chip that its
associated with.

This patch is required for nest IMC counters since, the counter values
are stored in an offset inside the HOMER region. To read those values
from the kernel for a chip, we need to associate the region to its
associated chip.

Signed-off-by: Vasant Hegde <hegdevasant@linux.vnet.ibm.com>
Signed-off-by: Hemant Kumar <hemant@linux.vnet.ibm.com>
Signed-off-by: Madhavan Srinivasan <maddy@linux.vnet.ibm.com>